requirements.txt
alembic==1.13.1
aniso8601==6.0.0
astroid==2.15.8
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Migrate==2.4.0
Flask-Moment==1.0.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.5.0
future==0.17.1
greenlet==3.0.3
importlib_metadata==7.1.0
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.0
Mako==1.3.2
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.9.9
psycopg2-binary==2.8.2
pyasn1==0.6.0
pycparser==2.21
pycryptodome==3.3.1
pylint==2.3.1
python-dotenv==1.0.1
python-jose==3.3.0
python-jose-cryptodome==1.3.2
pytz==2019.1
rsa==4.9
six==1.12.0
SQLAlchemy==1.4.52
typed-ast==1.4.2
typing_extensions==4.10.0
Werkzeug==0.15.6
wrapt==1.11.1
zipp==3.18.1
